Updated `DependenciesAwareTraitTest`
